Refactor (config): moved docstring from config to init
