+++ brevitas/nn/quant_linear.py
@@ -54,7 +54,7 @@ from brevitas.core.stats import StatsOp
 from brevitas.function.ops_ste import ceil_ste
 from brevitas.function.ops import max_uint
 from brevitas.proxy.parameter_quant import WeightQuantProxy, BiasQuantProxy, WeightReg
-from brevitas.config import docstrings
+from brevitas import docstrings
 from .quant_layer import QuantLayer, SCALING_MIN_VAL
 
 __all__ = ['QuantLinear']
